change pris to type int

+++ PulterkammeretShop/PulterkammeretShop/Models/Spil.cs
@@ -7,7 +7,7 @@ public class Spil
 {
     public int id { get; set; }
     public string navn { get; set; }
-    public double pris { get; set; }
+    public int pris { get; set; }
     public string kategori  { get; set; }
     public int antal { get; set; }
     
@@ -16,7 +16,7 @@ public class Spil
     ///
     /// skrevet af Kasper Sørensen.
     /// </summary>
-    public Spil(int id, string navn, double pris, string kategori)
+    public Spil(int id, string navn, int pris, string kategori)
     {
         this.id = id;
         this.navn = navn;
